Completed

+++ so_floodfill.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   so_floodfill.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: hulim <hulim@student.42singapore.sg>       +#+  +:+       +#+        */
+/*   By: hulim <hulim@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/14 20:45:59 by hulim             #+#    #+#             */
-/*   Updated: 2024/05/14 21:13:15 by hulim            ###   ########.fr       */
+/*   Updated: 2024/05/15 16:11:27 by hulim            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -43,9 +43,10 @@ char	**duplicatemap(t_game *mlxstruct)
 	dupmap = ft_calloc(getmapheight(mlxstruct->map) + 1, sizeof(char *));
 	while (i < getmapheight(mlxstruct->map))
 	{
-		dupmap[i] = ft_calloc(ft_strlen(mlxstruct->map[i]) + 1, sizeof(char));
+		dupmap[i] = ft_calloc((int)ft_strlen(mlxstruct->map[i]) + 1,
+				sizeof(char));
 		j = 0;
-		while (j < ft_strlen(mlxstruct->map[i]))
+		while (j < (int)ft_strlen(mlxstruct->map[i]))
 		{
 			dupmap[i][j] = mlxstruct->map[i][j];
 			j++;
@@ -57,7 +58,7 @@ char	**duplicatemap(t_game *mlxstruct)
 
 void	floodfill(char **dupmap, t_floodhelper *floodhelper, int x, int y)
 {
-	if (x < 0 || y < 0 || x >= ft_strlen(dupmap[y])
+	if (x < 0 || y < 0 || x >= (int)ft_strlen(dupmap[y])
 		|| y >= getmapheight(dupmap))
 		return ;
 	if (dupmap[y][x] == '1')
@@ -73,6 +74,6 @@ void	floodfill(char **dupmap, t_floodhelper *floodhelper, int x, int y)
 		floodfill(dupmap, floodhelper, x, y + 1);
 	if (x > 0)
 		floodfill(dupmap, floodhelper, x - 1, y);
-	if (x < ft_strlen(dupmap[y]) - 1)
+	if (x < (int)ft_strlen(dupmap[y]) - 1)
 		floodfill(dupmap, floodhelper, x + 1, y);
 }
